pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "studenttmixture"
readme = "README.md"
version = "1.11"
authors = [
    {name = "Jonathan Parkinson", email = "jlparkinson1@gmail.com"},
]
description = "Mixture modeling algorithms using the Student's t-distribution"
requires-python = ">=3.8"
keywords = ["mixture models", "T distribution", "student's T distribution",
    "mixture of T distributions", "mixture of Student's T distributions",
    "variational mixture", "variational inference"]
license = {file = "LICENSE"}
dependencies = [
    "numpy",
    "scipy",
    "scikit-learn"
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x"]

[project.urls]
Homepage = "https://github.com/jlparkI/mix_T"
Documentation = "https://github.com/jlparkI/mix_T"

[dependency-groups]
dev = [
    "hypothesis>=6.31.6",
    "jupyter>=1.1.1",
    "matplotlib>=3.3.4",
    "nbconvert>=6.0.7",
]
